Fix minor errors
